requirements.txt
pytest
selenium
page_objects
allure-pytest
mysql-connector-python